# docs/swagger.yaml
basePath: /
definitions:
  apierr.APIError:
    properties:
      detail:
        type: string
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
  response.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
...
  users.CreateUserRequest:
    properties:
      email:
        type: string
      is_admin:
        type: boolean
      password:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  users.UpdateUserRequest:
    properties:
      is_admin:
        type: boolean
      password:
        maxLength: 72
        minLength: 8
        type: string
    type: object
  users.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_admin:
        type: boolean
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: felipe@example.com
    name: Felipe
  description: REST API con Go + Gin · Auth JWT · PostgreSQL · Swagger
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: dev-test-api
  version: "1.0"
paths:
...
  /api/v1/profile:
    get:
      description: Obtiene los datos del usuario autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.APIError'
      security:
      - BearerAuth: []
      summary: Obtener perfil
      tags:
      - profile
...
  /api/v1/users:
    get:
      description: Lista todos los usuarios
      parameters:
      - description: 'Número de página (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Elementos por página (default: 20, max: 100)'
        in: query
        name: per_page
        type: integer
      - description: 'Campo de ordenación: email, created_at, updated_at'
        in: query
        name: sort_by
        type: string
      - description: 'Dirección: asc o desc (default: desc)'
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de usuarios (con paginación)
          schema:
            $ref: '#/definitions/response.Meta'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierr.APIError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/apierr.APIError'
      security:
      - BearerAuth: []
      summary: Listar usuarios (con paginación) (Admin)
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Crea un nuevo usuario
      parameters:
      - description: Datos del usuario
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/users.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierr.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierr.APIError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apierr.APIError'
      security:
      - BearerAuth: []
      summary: Crear usuario (Admin)
      tags:
      - users
  /api/v1/users/{id}:
    delete:
      description: Soft-delete de un usuario
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierr.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierr.APIError'
      security:
      - BearerAuth: []
      summary: Eliminar usuario (Admin)
      tags:
      - users
    get:
      description: Obtiene un usuario por ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierr.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierr.APIError'
      security:
      - BearerAuth: []
      summary: Obtener usuario (Admin)
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Actualiza los datos de un usuario
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Datos a actualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierr.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierr.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierr.APIError'
      security:
      - BearerAuth: []
      summary: Actualizar usuario (Admin)
      tags:
      - users
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'JWT token. Prefix with "Bearer ". Example: "Bearer eyJhbG..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"